requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
requests==2.31.0
urllib3~=2.4.0
pip~=25.1.1
redis~=5.2.1
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.8
Jinja2~=3.1.6
blinker~=1.9.0
itsdangerous~=2.2.0
certifi~=2025.1.31
idna~=3.10
charset-normalizer~=3.4.1
dotenv~=0.9.9
PyJWT~=2.10.1